search for tests in tests repo only
When you create a file with a name starting by test in a subdirectory of the tests it is found by run_tests.sh. Even if your file is not a test file, it is treated as one and this crashes the tests. This commit resolves the problem by searching for tests file only in the directory containing run_tests.